+++ apps/api/src/index.ts
@@ -1,9 +1,9 @@
 import "./instrument";
 
-import type { TypeBoxTypeProvider } from "@fastify/type-provider-typebox";
 import { PrismaClient } from "@prisma/client";
 import * as Sentry from "@sentry/node";
 import { Engine } from "@thirdweb-dev/engine";
+import { createAuth } from "@treasure-dev/auth";
 import { TREASURE_RUBY_CHAIN_DEFINITION } from "@treasure-dev/tdk-core";
 import { http, createConfig, fallback } from "@wagmi/core";
 import {
@@ -12,9 +12,8 @@ import {
   mainnet,
   sepolia,
 } from "@wagmi/core/chains";
-import Fastify from "fastify";
 import { createThirdwebClient } from "thirdweb";
-import { createAuth } from "thirdweb/auth";
+import { createAuth as createThirdwebAuth } from "thirdweb/auth";
 import { privateKeyToAccount } from "thirdweb/wallets";
 import { defineChain } from "viem";
 
@@ -30,13 +29,16 @@ import { projectsRoutes } from "./routes/projects";
 import { transactionsRoutes } from "./routes/transactions";
 import { usersRoutes } from "./routes/users";
 import type { TdkApiContext } from "./types";
+import { app } from "./utils/app";
 import { getEnv } from "./utils/env";
 
 const main = async () => {
-  const app = Fastify().withTypeProvider<TypeBoxTypeProvider>();
-
   const env = await getEnv();
   const client = createThirdwebClient({ secretKey: env.THIRDWEB_SECRET_KEY });
+  const adminAccount = privateKeyToAccount({
+    client,
+    privateKey: env.THIRDWEB_AUTH_PRIVATE_KEY,
+  });
   const ctx: TdkApiContext = {
     env,
     db: new PrismaClient({
@@ -48,12 +50,15 @@ const main = async () => {
     }),
     client,
     auth: createAuth({
-      domain: env.THIRDWEB_AUTH_DOMAIN!,
+      kmsKey: env.TREASURE_AUTH_KMS_KEY,
+      issuer: env.THIRDWEB_AUTH_DOMAIN,
+      audience: adminAccount.address,
+      expirationTimeSeconds: 86_400, // 1 day
+    }),
+    thirdwebAuth: createThirdwebAuth({
+      domain: env.THIRDWEB_AUTH_DOMAIN,
       client,
-      adminAccount: privateKeyToAccount({
-        client,
-        privateKey: env.THIRDWEB_AUTH_PRIVATE_KEY!,
-      }),
+      adminAccount,
       login: {
         uri: `https://${env.THIRDWEB_AUTH_DOMAIN}`,
       },
@@ -107,21 +112,19 @@ const main = async () => {
 
   // Middleware
   Sentry.setupFastifyErrorHandler(app);
-  await withSwagger(app);
-  await withCors(app);
-  await withErrorHandler(app);
-  await withChain(app);
-  await withAuth(app, ctx);
+  withCors(app);
+  withSwagger(app);
+  withErrorHandler(app);
+  withChain(app);
+  withAuth(app, ctx);
 
   // Routes
-  await Promise.all([
-    app.register(authRoutes(ctx)),
-    app.register(usersRoutes(ctx)),
-    app.register(projectsRoutes(ctx)),
-    app.register(transactionsRoutes(ctx)),
-    app.register(harvestersRoutes(ctx)),
-    app.register(magicswapRoutes(ctx)),
-  ]);
+  app.register(authRoutes(ctx));
+  app.register(usersRoutes(ctx));
+  app.register(projectsRoutes(ctx));
+  app.register(transactionsRoutes(ctx));
+  app.register(harvestersRoutes(ctx));
+  app.register(magicswapRoutes(ctx));
 
   app.get("/healthcheck", async (_, reply) => {
     try {
@@ -146,13 +149,11 @@ const main = async () => {
 
   // Start server
   await app.ready();
-  app.listen({ host: "0.0.0.0", port: Number(env.PORT) }, (err, address) => {
+  app.listen({ host: "0.0.0.0", port: Number(env.PORT) }, (err) => {
     if (err) {
-      console.error("Error starting server:", err);
+      app.log.error("Error starting server:", err);
       process.exit(1);
     }
-
-    console.log(`Server listening at ${address}`);
   });
 };