+++ apps/api/src/schema/auth.ts
@@ -42,6 +42,8 @@ export const loginReplySchema = Type.Object({
     userSchema,
     userProfileSchema,
     Type.Object({
+      sessions: Type.Array(sessionSchema),
+      // Fields for backwards compatibility
       allActiveSigners: Type.Array(sessionSchema),
     }),
   ]),