+++ packages/core/src/index.ts
@@ -32,7 +32,7 @@ export {
 } from "./utils/date";
 
 // Connect
-export { getAllActiveSigners } from "./connect/accounts";
+export { generateAccountSignature } from "./connect/auth";
 export { decodeAuthToken } from "./connect/jwt";
 export {
   type SocialConnectMethod,
@@ -48,8 +48,9 @@ export {
 } from "./connect/login";
 export {
   createSession,
-  validateSession,
+  getUserSessions,
   startUserSession,
+  validateSession,
 } from "./connect/session";
 
 // Bridgeworld