+++ packages/react/src/components/connect/ConnectMethodSelectionView.tsx
@@ -12,6 +12,7 @@ import { GoogleIcon } from "../../icons/GoogleIcon";
 import { PasskeyIcon } from "../../icons/PasskeyIcon";
 import { TelegramIcon } from "../../icons/TelegramIcon";
 import { WalletIcon } from "../../icons/WalletIcon";
+import { XIcon } from "../../icons/XIcon";
 import { cn } from "../../utils/classnames";
 import { Button } from "../ui/Button";
 import { ConnectFooter } from "./ConnectFooter";
@@ -107,57 +108,62 @@ export const ConnectMethodSelectionView = ({
           {t("common.or")}
         </span>
       </div>
-      <div
-        className={cn(
-          "tdk-grid tdk-gap-2 tdk-grid-cols-2",
-          disablePasskey ? "md:tdk-grid-cols-4" : "md:tdk-grid-cols-5",
-        )}
-      >
+      <div className="tdk-grid tdk-gap-2 tdk-grid-cols-6 md:tdk-grid-cols-5">
         <ConnectMethodButton
+          className="tdk-col-span-2 md:tdk-col-span-1"
+          title={t("connect.option.google")}
           onClick={() => onConnect("google")}
           disabled={isLoading}
         >
-          <GoogleIcon className="tdk-w-6 tdk-h-6 tdk-mx-auto" />
-          <span className="tdk-block">{t("connect.option.google")}</span>
+          <GoogleIcon className="tdk-w-7 tdk-h-7" />
         </ConnectMethodButton>
         <ConnectMethodButton
+          className="tdk-col-span-2 md:tdk-col-span-1"
+          title={t("connect.option.x")}
+          onClick={() => onConnect("x")}
+          disabled={isLoading}
+        >
+          <XIcon className="tdk-w-5 tdk-h-5" />
+        </ConnectMethodButton>
+        <ConnectMethodButton
+          className="tdk-col-span-2 md:tdk-col-span-1"
+          title={t("connect.option.telegram")}
           onClick={() => onConnect("telegram")}
           disabled={isLoading}
         >
-          <TelegramIcon className="tdk-w-6 tdk-h-6 tdk-mx-auto" />
-          <span className="tdk-block">{t("connect.option.telegram")}</span>
+          <TelegramIcon className="tdk-w-6 tdk-h-6" />
         </ConnectMethodButton>
+        <span className="md:tdk-hidden" />
         <ConnectMethodButton
+          className="tdk-col-span-2 md:tdk-col-span-1"
+          title={t("connect.option.discord")}
           onClick={() => onConnect("discord")}
           disabled={isLoading}
         >
-          <DiscordIcon className="tdk-w-6 tdk-h-6 tdk-mx-auto" />
-          <span className="tdk-block">{t("connect.option.discord")}</span>
+          <DiscordIcon className="tdk-w-6 tdk-h-6" />
         </ConnectMethodButton>
         <ConnectMethodButton
-          className="tdk-space-y-0.5"
+          className="tdk-col-span-2 md:tdk-col-span-1"
+          title={t("connect.option.apple")}
           onClick={() => onConnect("apple")}
           disabled={isLoading}
         >
-          <AppleIcon className="tdk-w-5 tdk-h-5 tdk-mx-auto" />
-          <span className="tdk-block">{t("connect.option.apple")}</span>
+          <AppleIcon className="tdk-w-5 tdk-h-5" />
         </ConnectMethodButton>
+        <span className="md:tdk-hidden" />
         {!disablePasskey ? (
           <ConnectMethodButton
-            className="tdk-col-span-2 md:tdk-col-span-1 tdk-flex tdk-items-center tdk-justify-center tdk-gap-1 tdk-py-2 md:tdk-block"
+            className="tdk-flex tdk-items-center tdk-gap-1 tdk-justify-center tdk-py-2 tdk-col-span-full"
             onClick={() => onConnect("passkey")}
             disabled={isLoading}
           >
-            <PasskeyIcon className="tdk-w-6 tdk-h-6 md:tdk-mx-auto" />
+            <PasskeyIcon className="tdk-w-6 tdk-h-6" />
             <span className="tdk-block">{t("connect.option.passkey")}</span>
           </ConnectMethodButton>
         ) : null}
         {!disableWallet ? (
           <ConnectMethodButton
-            className={cn(
-              "tdk-flex tdk-items-center tdk-gap-1 tdk-justify-center tdk-py-2 tdk-col-span-2",
-              disablePasskey ? "md:tdk-col-span-4" : "md:tdk-col-span-5",
-            )}
+            className="tdk-flex tdk-items-center tdk-gap-1 tdk-justify-center tdk-py-2 tdk-col-span-full"
             onClick={() => onConnect("wallet")}
             disabled={isLoading}
           >
@@ -179,7 +185,7 @@ const ConnectMethodButton = ({
     <button
       type="button"
       className={cn(
-        "tdk-bg-night-500 tdk-border tdk-border-solid tdk-border-night-400 tdk-p-3 tdk-text-xs tdk-text-silver-100 tdk-font-medium tdk-transition-colors tdk-rounded-lg",
+        "tdk-flex tdk-items-center tdk-justify-center tdk-bg-night-500 tdk-border tdk-border-solid tdk-border-night-400 tdk-p-3 tdk-text-xs tdk-text-silver-100 tdk-font-medium tdk-transition-colors tdk-rounded-lg",
         props.disabled
           ? "tdk-opacity-50 tdk-cursor-not-allowed"
           : "tdk-cursor-pointer hover:tdk-bg-cream hover:tdk-border-cream hover:tdk-text-night-800",